% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numverify}
\alias{numverify}
\title{Numverify : Validate Phone Number}
\usage{
numverify(country_code = "CN", number = NULL)
}
\arguments{
\item{country_code}{Country code, default China : "CN".}

\item{number}{Phone number}
}
\value{
the results from the validation
}
\description{
Numverify : Validate Phone Number
}
\examples{
\dontrun{
numverify(number = 14158586273)
}
}
